Create Dashboard.jsx
Move Dashboard.jsx to BrowseProjects Folder
